Version 0.77a2 :
- The debugger starts to work again (with the new modelisation)

        . FIFO communication between client and server works
        . Allocation proxying works
        . Linkmap symbol resolving works

- Right now there is a bug in debug format loading from the debugger, trying
to fix it. This happens when loading the files at the beginning.

- Now we sometimes have to compile librevm twice (when readline is enabled)
because we also need a non-readline version of it for the embedded debugger

- Now the debugger is a single library (libe2dbg.so), all .o of the framework
are linked in a single lib, so that we have less dependencies intrusion in the
host process. I had to change all Makefiles for this, so that libraries also
generates their own libname.o that contains all the code of it.

- The debugger is now a real C program that stands in e2dbg/e2dbg. It
automatically LD_PRELOAD the libe2dbg in the target process and fork it.

=> Continuing the work. Need to make the breakpoints and stepping to work
on both single thread and multithreads programs.

-may