- The debugger works on multithread !
It never worked as good. I developped and tested on linux. It works
perfectly on both monothread and multithread processes.

To be done :

- there is still a little prompt synchronization problem between
the client and server part of the debugger. This is not a big deal I will
correct it asap.

- debugger scripting is untested on e2dbg -current

- debugging ET_DYN process is untested on e2dbg -current

- ET_REL injection in memory is untested on e2dbg -current

Strauss: you mentioned that feature would be cool, actually we already have
it since the "Embedded ELF Debugging" phrack article, almost 2 years. All
ET_REL injection and function redirection works in memory since that time
(version 0.65)

Priority TODO:

- Make the multithread embedded debugger to work and be tested on BSD.

Im starting to do that now.

-may