New branch.
This branch will be used to modify the build process.
configure will not generate the makefile anymore and
rules will be separated in mltiple sub makefiles.

This is intended to gain in clarity as eresi is beginning
to have several build configurations, several needs.

Any suggestion, comments or idea are welcomed on this branch.
Do not hesitate to report any problem.