summary: Testsuite compilation fix.
Some testuite directories were not being linked
with etrace.

Note about this:

It sounds strange to me to need to link with
libetrace just to use elfsh features like in
ctors_hijack.

If libelfsh is depending on libetrace, this should
be corrected.

The env.sh script is also adding to PATH
the directories of elfsh, etrace and kernsh.